Move predefined for properties into own namespace and add metadata
