+++ src/Models/Metadata/Predefined/PredefinedResolverInterface.php
@@ -1,0 +1,28 @@
+<?php
+declare(strict_types=1);
+
+/**
+ * Pimcore
+ *
+ * This source file is available under two different licenses:
+ * - GNU General Public License version 3 (GPLv3)
+ * - Pimcore Commercial License (PCL)
+ * Full copyright and license information is available in
+ * LICENSE.md which is distributed with this source code.
+ *
+ *  @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
+ *  @license    http://www.pimcore.org/license     GPLv3 and PCL
+ */
+
+namespace Pimcore\Bundle\StaticResolverBundle\Models\Metadata\Predefined;
+
+use Pimcore\Model\Metadata\Predefined;
+
+interface PredefinedResolverInterface
+{
+    public function getById(string $id): ?Predefined;
+
+    public function getByName(string $name): ?Predefined;
+
+    public function create(): Predefined;
+}
